+++ CHANGELOG.md
@@ -4,4 +4,5 @@
 
 ## 0.1.0 March 5, 2024
 
+- Enh #2: Add helper `Template::class` (@terabytesoftw)
 - Initial release